chore(server): refactor server code
