+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "nyc mocha -r @babel/register --exit",
-    "start": "node -r @babel/register server/server",
+    "start": "nodemon -r @babel/register server/server",
     "coverage": "nyc report --reporter=test-lcov | coveralls"
   },
   "repository": {
@@ -20,10 +20,10 @@
   "homepage": "https://github.com/boscoseries/iReporter#readme",
   "dependencies": {
     "body-parser": "^1.18.3",
+    "coveralls": "^3.0.2",
     "express": "^4.16.4",
     "ghooks": "^2.0.4",
-    "nyc": "^13.1.0",
-    "request": "^2.88.0"
+    "nyc": "^13.1.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.2.0",
